final refactoring and logging improvements
